move div table styles to common styles
